vrclient: Make IVRApplications::LaunchInternalProcess() manual on the PE side.
